Finishing the get computers choice funtion.
I created a funtion called getComputersChoice().
The purpose of this function was to randomly
choose between rock, paper, or scissors. We didn't
have to add any arguments to the funtion because
it didn't require any specific input. Now we had
to add a lot to the body of the we added a
constant variable named randomNumber and asigned
it the Math object with a method of .random().
After this we created a if statement. The if
statement said if the randomNumber was less than
0.3 return "rock", if else the randomNumber was less
than 0.6 then return "paper", and else the
randomNumber was higher than 0.6 return scissors.
All of this was returned in the console with
console.log(getComputersChoice());.