Finishing the getHumanChoice() function.
First I created a function called getHumanChoice()
We didn't need any arguments because it didn't
require any inputs. I created a a constant
variable and named it answer and asigned it a
prompt that says "rock, paper, or scissors?"
asking the player what they want to choose. To
display the answer in the console I created a
switch statment. I connected the answer exspression
to the switch statment switch(answer). They I
made if so if the player typed in rock, it would
return rock, and did the same with paper, and scissors.
Then I console.loged it with console.log(getHumanChoice());.